requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
python-multipart==0.0.9
transformers==4.42.4
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1